#13_Added the necessary CI tests, and also transferred functions from cpp to hpp.
