+++ .github/workflows/ci.yml
@@ -1,0 +1,28 @@
+name: Gomoku Bot CI
+
+on: [push, pull_request]
+
+jobs:
+  build-and-test:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Install Dependencies
+        run: |
+          sudo apt-get update
+          sudo apt-get install -y cmake g++ libgtest-dev
+
+      - name: Configure CMake
+        run: cmake -B build -S . -DBUILD_TTTCORE=FULL
+
+      - name: Build
+        run: cmake --build build --config Release
+
+      - name: Run My Tests
+        run: |
+          cd build
+          # запуск только моих тестов
+          ctest --output-on-failure -R MyBotLogicTests
